+++ src/main/scala/xitrum/routing/SockJsRouteMap.scala
@@ -56,7 +56,7 @@ class SockJsRouteMap(map: MMap[String, SockJsClassAndOptions]) extends Log {
   def createSockJsAction(context: ActorRefFactory, pathPrefix: String): ActorRef = {
     val sockJsClassAndOptions = map(pathPrefix)
     val actorClass            = sockJsClassAndOptions.actorClass
-    context.actorOf(Props(Dispatcher.newActionInstance(actorClass).asInstanceOf[Actor]))
+    context.actorOf(Props(Dispatcher.newAction(actorClass).asInstanceOf[Actor]))
   }
 
   /** @param sockJsHandlerClass Normal SockJsHandler subclass or object class */